add fetch all
