+++ app.js
@@ -1,12 +1,14 @@
-const express = require('express')
-const logger = require('morgan')
-const cors = require('cors')
+const express = require('express');
+const logger = require('morgan');
+const cors = require('cors');
+const dotenv = require('dotenv')
+dotenv.config();
+
+global.basedir = __dirname;
 
 const contactsRouter = require('./routes/api/contacts')
-
 const app = express()
-
-const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'
+const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short';
 
 app.use(logger(formatsLogger))
 app.use(cors())
@@ -19,7 +21,7 @@ app.use((req, res) => {
 })
 
 app.use((err, req, res, next) => {
-  const { status = 500, message = 'Server error' } = err;
+  const { status = 500, message = 'Internal Server error' } = err;
   res.status(status).json({ message })
 })
