+++ controllers/contacts/getAllContacts.js
@@ -1,0 +1,16 @@
+const { basedir } = global;
+const { Contact } = require(`${basedir}/models/contact`);
+
+
+const getAllContacts = async (req, res, next) => {
+    // try {
+    console.log("global ",basedir);
+    const result = await Contact.find();
+    res.json(result)
+    // } catch (error) {
+    //   next(error);
+    // }
+};
+
+
+module.exports = getAllContacts;
