+++ controllers/contacts/index.js
@@ -1,0 +1,7 @@
+const getAllContacts = require('./getAllContacts');
+
+
+module.exports = {
+    getAllContacts,
+    
+}
