+++ helpers/index.js
@@ -1,5 +1,7 @@
 const createError = require('./createError');
+const ctrlWrapper = require('./ctrlWrapper');
 
 module.exports = {
     createError,
+    ctrlWrapper,
 }
