+++ schemas/index.js
@@ -1,5 +1,5 @@
-const contactSchema = require('./contactSchema');
+// const contactSchema = require('./contactSchema');
 
-module.exports = {
-    contactSchema,
-}
+// module.exports = {
+//     contactSchema,
+// }
