+++ server.js
@@ -3,6 +3,7 @@
 //   console.log("Server running. Use our API on port: 3000")
 // })
 
+
 const app = require('./app');
 const { default: mongoose } = require('mongoose');
 const { DB_HOST, PORT } = process.env;
